country iso

+++ migrations/V0014__config_taxonomies.sql
@@ -7,11 +7,11 @@
 
 -- categories table (hierarchical) - “What is this post generally about?”
 create table config_dynamic_01_taxonomy_categories (
-    id bigint primary key,
+    id int8 primary key,
     name varchar(100) not null,
     slug varchar(120) not null unique,
     description text,
-    parent_id bigint references config_dynamic_01_taxonomy_categories(id) on delete cascade,
+    parent_id int8 references config_dynamic_01_taxonomy_categories(id) on delete cascade,
     color varchar(7) default '#3b82f6',
     icon varchar(50),
     
@@ -27,7 +27,7 @@ create table config_dynamic_01_taxonomy_categories (
     is_visible boolean not null default true,
     
     -- metadata
-    updated_at bigint,
+    updated_at int8,
     
     -- path for hierarchical queries
     path ltree -- postgresql ltree extension for hierarchical queries
@@ -39,7 +39,7 @@ create index idx_categories_path on config_dynamic_01_taxonomy_categories using 
 
 -- tags table - “What topics, tools, or ideas appear in this post?”
 create table config_dynamic_02_taxonomy_tags (
-    id bigint primary key,
+    id int8 primary key,
     name varchar(100) not null,
     slug varchar(120) not null unique,
     description text,
@@ -52,7 +52,7 @@ create table config_dynamic_02_taxonomy_tags (
     post_count integer not null default 0,
     
     -- metadata
-    updated_at bigint
+    updated_at int8
 );
 
 create index idx_tags_slug on config_dynamic_02_taxonomy_tags(slug);
